split WeaponHandler - WeaponConfig
